chart-incident-management/requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
pillow>=10.0.0
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
pydantic>=2.4.0
inference>=0.15.0
ultralytics>=8.4.0
transformers>=4.45.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
dash>=4.1.0
dash-bootstrap-components>=2.0.4
plotly>=6.6.0
reportlab>=4.0.0